+++ packages/website/pages/examples/[...example].js
@@ -1,3 +1,4 @@
+import { Suspense } from "react";
 import { useRouter } from "next/router";
 import CommonHead from "/components/CommonHead";
 import { table, components } from "/AllExamples";
@@ -14,9 +15,14 @@ const Example = () => {
             <h1 className="text-4xl lg:col-span-2 mt-1 mb-2 font-bold text-center ml-14 sm:ml-0">
                 {title}
             </h1>
-            <Component />
+            <Suspense>
+                <Component />
+            </Suspense>
+
             <pre className="language-jsx overflow-auto !m-0">
-                <ComponentCode />
+                <Suspense>
+                    <ComponentCode />
+                </Suspense>
             </pre>
         </div>
     );