+++ packages/af-utils-react-virtual-list/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/react-virtual-list
 
+## 0.0.10
+
+### Patch Changes
+
+-   added react 18 support
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.9
+
 ## 0.0.9
 
 ### Patch Changes